cosmic-cove/python/requirements.txt
strawberry-graphql
fastapi
uvicorn
httpx